TRAINING SET CODES AND RESULTS - NLP MULTI-CLASS CLASSIFICATION CASE STUDY
This is a Python file that contains the results, inferences and codes applied to the Multi-class classification problem of building a Multi-Nomial Naive Bayes classifier to classify customer text messages into 10 different label categories. The codes, inferences and results pertain to the Exploratory Data Analysis stage, parameter tuning stage and the model building stage.